netsniff-ng: Fix compiler warning
Fix the following compiler warning that occurs when building with "-W
-Wall -Wextra" by introducing a cast:

netsniff-ng.c: In function ‘walk_t3_block’:
netsniff-ng.c:841:8: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ netsniff-ng.c
@@ -838,7 +838,7 @@ static void walk_t3_block(struct block_desc *pbd, struct ctx *ctx,
 
 			ret = __pcap_io->write_pcap(*fd, &phdr, ctx->magic, packet,
 						    pcap_get_length(&phdr, ctx->magic));
-			if (unlikely(ret != pcap_get_total_length(&phdr, ctx->magic)))
+			if (unlikely(ret != (int) pcap_get_total_length(&phdr, ctx->magic)))
 				panic("Write error to pcap!\n");
 		}
